bloglist expansion, step9

+++ part4/blog-list/controllers/blogs.js
@@ -41,8 +41,26 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete( '/:id', async ( request, response, next ) => {
 	try {
-		await Blog.findByIdAndRemove( request.params.id )
-		response.status( 204 ).end()
+		const decodedToken = jwt.verify(request.token, process.env.SECRET)
+		if (!decodedToken.id) {
+			return response.status(401).json({ error: 'token invalid' })
+		}
+
+		const user = await User.findById( decodedToken.id )	
+		const blog = await Blog.findById( request.params.id )
+		
+		console.log('blog user', blog.user._id.toString());
+		console.log( 'user from token', user._id.toString() )
+		if ( blog.user._id.toString() === user._id.toString() ) {
+			
+			await Blog.findByIdAndRemove( request.params.id )
+			response.status( 204 ).end()
+		} else {
+			response.status( 401 ).json( {
+				error: 'unauthorized user'
+			})
+		}
+		
 	} catch (error) {
 		next(error)
 	}
